feat: add shared relative time formatter
